put login button to navbar (#454)
* toggle between sign in and logout

* styling the sign in button and removing unused comment

* changed the logout button position
