+++ assets/css/app.scss
@@ -110,6 +110,8 @@
   margin: 0;
   padding: 0;
   background-color: #333;
+  align-items: center;
+  display: flex;
 }
 
 .nav li {
